+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/settings/ProjectSettings.kt
@@ -22,6 +22,7 @@ class ProjectSettings : PersistentStateComponent<ProjectSettings?> {
     }
 
     var gitlabUrl: String? = null
+    var gitlabUrls: Set<String> = setOf()
     var fallbackBranch = "main"
     var remote = "origin"
     var forceHttps = true